+++ lib/h2.mli
@@ -65,12 +65,10 @@ module Method : module type of Httpaf.Method
     See {{:https://tools.ietf.org/html/rfc7231#section-6} RFC7231§6} for more
     details.
 
-    For the most part, this module is a proxy to [Httpaf.Status]. Its
-    [informational] type, however, removes support for the [Switching_protocols]
-    tag, as per the HTTP/2 spec (relevant portion reproduced below).
-
-    From RFC7540§8.1.1: HTTP/2 removes support for the 101 (Switching
-    Protocols) informational status code ([RFC7231], Section 6.2.2).
+    This module is a strict superset of [Httpaf.Status]. Even though the HTTP/2
+    specification removes support for the [Switching_protocols] status code, h2
+    keeps it for the sake of higher level interaction between OCaml libraries
+    that support both HTTP/1 and HTTP/2.
 
     See {{:https://tools.ietf.org/html/rfc7540#section-8.1.1} RFC7540§8.1.1}
     for more details. *)